Adds info of other servers to Proxy Server + introduces concept of load
ProxyServer class now takes input about nunber of other servers and
the ID of this server.

ServerLoad.java:
A class to store load on a server. It includes a integer value of load
and a timestamp of when the value was last updated.
It also includes a method to return appropriate threshold for normal
state which is calculated as the average of loads on all servers.
The threshold for over load is assumed as 1.3 times threshold for
normal.