Adds Simultaneous Requesting + Bug Fixes
Modifies the LoopedRequester to make multiple simultaneous requests
parallely

Modifies IPFunctions to return a loopback address if non loopback
address is not found in getFirstNonLoopbackAddress(). This solves an
issue that was preventing the simulation from running on same machine
when not connected to a network. Although this spoils the semantics
of the function name.

Fixes a bug where the Server name label was not getting set in the
GUI.