Adds tracking of IP and port of other servers
The ServerLoad class now stores the IP and Port of the server it is
referring to. This is used by each proxy server to keep track of
the IP and Port of other proxy servers.

Adds listening for Load updates. If new informations is received,
loads are updated in the ServerLoad class.